refactor bird module
